+++ extension/library/indicators/portrait.js
@@ -1,0 +1,5 @@
+export {
+    ensureBatteryIndicator as ensurePortraitIndicator,
+    destroyBatteryIndicator as destroyPortraitIndicator,
+    updateBatteryIndicatorStatus as updatePortraitIndicatorStatus,
+} from './battery.js';